Updated the Read Me file

+++ hello_world/6-size.c
@@ -13,7 +13,7 @@ long long int c;
 char d;
 float e;
 printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of an int: %d byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of an int: %d byte(s)\n", (int)sizeof(a));
 printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
 printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
 printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
